expand options

+++ include/centrolign/utility.hpp
@@ -25,7 +25,7 @@ std::string decode_seq(const std::string& seq);
 // return pairs of (name, sequence) from a FASTA
 std::vector<std::pair<std::string, std::string>> parse_fasta(std::istream& in);
 
-int parse_int(const std::string& str);
+int64_t parse_int(const std::string& str);
 
 template<class BGraph>
 void print_graph(const BGraph& graph, std::ostream& out);
